Less warning for indent

+++ src/uncrustify_types.h
@@ -88,18 +88,18 @@ struct paren_stack_entry_t
 {
    c_token_t    type;         /**< the type that opened the entry */
    size_t       level;        /**< Level of opening type */
-   int          open_line;    /**< line that open symbol is on */
+   size_t       open_line;    /**< line that open symbol is on */
    chunk_t      *pc;          /**< Chunk that opened the level */
    int          brace_indent; /**< indent for braces - may not relate to indent */
-   int          indent;       /**< indent level (depends on use) */
+   size_t       indent;       /**< indent level (depends on use) */
    size_t       indent_tmp;   /**< temporary indent level (depends on use) */
-   int          indent_tab;   /**< the 'tab' indent (always <= real column) */
+   size_t       indent_tab;   /**< the 'tab' indent (always <= real column) */
    bool         indent_cont;  /**< indent_continue was applied */
    int          ref;
    c_token_t    parent;       /**< if, for, function, etc */
    brstage_e    stage;
    bool         in_preproc;   /**< whether this was created in a preprocessor */
-   int          ns_cnt;
+   size_t       ns_cnt;
    bool         non_vardef;   /**< Hit a non-vardef line */
    indent_ptr_t ip;
 };
